Applied Dan's fixes on the 19 century's problem
The code only processes the model outputs in 80's
and 90's for 19 century to save time. Dan Ricciuto
fixed it.

As there are only NEE and NBP that have observational
data earlier than 1980 for model comparison, the Dan's fix
will let the code deal with NEE and NBP data before 1980.